+++ changelog.md
@@ -1,3 +1,6 @@
+# Next
+* Added function "/vmod op prune-shipyard-chunks" to delete chunks of deleted ships
+
 # 1.7.1
 * Fixed bug with "Open or Close Toolgun GUI" keybind stopping working sometimes
 * Fixed ~~maybe~~ some issues with schematics